encryptedPassword done & registration method adds to the user.
