--- api/Register.api.js
@@ -1,15 +1,0 @@
-const {save} = require('../dal/Register.dao');
-
-const createUser = async ({fullName, email, type, password}) => {
-    const user ={
-        fullName,
-        email,
-        type,
-        password
-    }
-    return await save(user);
-}
-
-module.exports = {
-    createUser
-};
