+++ api/User.api.js
@@ -1,4 +1,16 @@
-const {getAll, getById, removeById, update} = require('../dal/User.dao');
+const {save, getAll, getById, removeById, update} = require('../dal/User.dao');
+const {encrypt} = require("../EncryptionHandler");
+
+const createUser = async ({fullName, email, type, password}) => {
+    const hashedPassword = encrypt(password);
+    const user ={
+        fullName,
+        email,
+        type,
+        password: hashedPassword
+    }
+    return await save(user);
+}
 
 const getAllUser = async () => {
     return await getAll();
@@ -13,16 +25,18 @@ const deleteUser = async (id) => {
 };
 
 const updateUser = async (id, {fullName, email, type, password}) => {
+    const hashedPassword = encrypt(password);
     const user ={
         fullName,
         email,
         type,
-        password
+        password: hashedPassword
     }
     return await update(id,user);
 };
 
 module.exports = {
+    createUser,
     getAllUser,
     getUserById,
     deleteUser,
