+++ dal/User.dao.js
@@ -1,5 +1,12 @@
 
 const Users = require('./').db('conferenceMT').collection('userCollection');
+const objectID = require("mongodb").ObjectId;
+
+const save = async ({fullName, email, type, password}) => {
+    const result = await Users.insertOne({fullName, email,
+        type, password});
+    return result.ops[0];
+};
 
 const getAll = async ()=> {
     const cursor = await Users.find();
@@ -7,11 +14,11 @@ const getAll = async ()=> {
 };
 
 const getById = async id => {
-    return await Users.findOne({id});
+    return await Users.findOne({_id:objectID(id)});
 };
 
 const removeById =async id => {
-    return await Users.deleteOne({id});
+    return await Users.deleteOne({_id:objectID(id)});
 };
 
 const update = async (id, {fullName, email, type, password}) => {
@@ -20,6 +27,7 @@ const update = async (id, {fullName, email, type, password}) => {
 };
 
 module.exports = {
+    save,
     getAll,
     getById,
     removeById,
